Add scoped variables and parameters

+++ Script.Core/Framework/Math/LessThan.cs
@@ -13,8 +13,8 @@
         /// <inheritdoc/>
         public override string Execute(ExecutionContext context, params string[] parameters)
         {
-            var a = Cast.ToInt(parameters[0]);
-            var b = Cast.ToInt(parameters[1]);
+            var a = Cast.ToDouble(parameters[0]);
+            var b = Cast.ToDouble(parameters[1]);
 
             return Cast.ToString(a < b);
         }
